80% done SSK Hud
Added SSK life counter, star counter, and coin counter, as well as the crosses/X's next to them.
Originally the X's were going to be a text class included with the quotation marks and apostrophe, except it did **not** want to work with me, so I instead made it extend FlxSprite. You can now use `("entityname", x, y, scale)` to configure it. Also did changes to the numbers class, fixing a bug I was having that crashed due to it trying to load the wrong image.

Todos:
- Add health meter.
- Add icons.
- Add health meter functionality.