Mostly Done HP Meter Movement (SSK)
Note there are a few flaws I can't fix. For instance, if you were to press down to stimulate getting damage from full HP, it'd normally go to an area to the middle, however, if you do this multiple times, it'd just go to where it'd just be at casually (close to the top of the screen but not being offscreen.)

If you recover all your health normally, the HP bar goes up offscreen after 2 seconds. If you do this multiple times, it goes up instantly.

I don't know how to fix these issues, sorry.